spaceController, service, dao, mapper 수정 - 공간 상세페이지 같은 카테고리 공간 조회

+++ SpaceUs/src/main/java/com/kh/spaceus/space/model/vo/Space.java
@@ -31,7 +31,7 @@ public class Space implements Serializable{
 	private int views;
 	private int likeCnt;
 	private String status;
-	private int account;
+	private long account;
 	private String content;
 	private String bank;
